Solve Day 9 part 1 with Clash, answered with real hardware but with a pretty major hack.
Here, the code has to be recompiled and run four different times,
calculating only the squares visited in a single quadrant at a time.
This is a solveable problem (we can buffer parsed input, and replay it
four times), but this work just isn't all that interesting, especially
since we're not _that_ far off of having enough memory.  We also still
don't yet know what part 2 offers, and it may offer yet another major
shake up.