Solve 2022 Day 9 part 2 with Clash, on real hardware, but with one major hack.
We continue here to recompile and rerun per quadrant (acually, I
realized it's a stripe later, nbd).  The only way to make this work
would be to buffer the whole input and re-execute on the chip per
stripe/quadrant.  Since the dedup is even unique by part, we would
actually need to run the input through 8 times!  Again, this part just
isn't that interesting to actually code.

It's a shame because this has been _the best_ example of what custom
hardware can do, outside of the memory limits, because we can make an
enormously long rope with only a latency penalty with pipelining.